refactor: 떡국 완성 함수 위치 변경
- 로그인 페이지 가운데 정렬 및 이미지 위치 수정
- 이용약관에 underline 및 href
- 떡국 완성 함수 위치 변경
